声音全部改为8位无符号，因为实际分辨率无法达到这么高，速度不够

+++ oscilloscope.h
@@ -12,8 +12,8 @@
 #include <QAudioFormat>
 
 struct Point {
-    qint16 x = 0;
-    qint16 y = 0;
+    quint8 x = 0;
+    quint8 y = 0;
 };
 
 class Oscilloscope : public QThread
